requirements.txt
aiofiles~=23.2.1
httpx~=0.24.1
pyuseragents~=1.0.5
numpy~=1.26.2
loguru~=0.7.2
beautifulsoup4~=4.12.2
aiocsv~=1.2.5
PyYAML~=6.0.1
postgrest~=0.13.0
supabase~=2.1.0